Move MethodType static helpers to new helper Class
static methods:
- vmResolveFromMethodDescriptorString()
- parseIntoClass()
- nonPrimitiveClassFromString()
- getBytecodeStringName()
- unwrapPrimitive()
- wrapPrimitive()

static fields:
- primitivesArray
- WRAPPER_SET
