+++ jcl/src/java.base/share/classes/java/lang/Class.java
@@ -4842,7 +4842,7 @@ SecurityException {
 	 * @return field descriptor of Class instance
 	 */
 	public String descriptorString() {
-		/* see MethodType.getBytecodeStringName */
+		/* see MethodTypeHelper.getBytecodeStringName */
 
 		if (this.isPrimitive()) {
 			if (this == int.class) {
