+++ jcl/src/java.base/share/classes/java/lang/invoke/MethodType.java
@@ -75,33 +75,11 @@ public final class MethodType implements Serializable
 /*[ENDIF]*/
 {
 	static final Class<?>[] EMTPY_PARAMS = new Class<?>[0];
-	static final Set<Class<?>> WRAPPER_SET;
-	static {
-		Class<?>[] wrappers = {Byte.class, Character.class, Double.class, Float.class, Integer.class, Long.class, Short.class, Boolean.class };
-		WRAPPER_SET = Collections.unmodifiableSet(new HashSet<Class<?>>(Arrays.asList(wrappers)));
-	}
 
 /*[IF Sidecar19-SE-OpenJ9]*/	
 	private MethodTypeForm form;
 /*[ENDIF]*/	
-	
-	/*[IF ]*/
-	/* Do not include 'primitives.put("V", void.class)' as void.class is not yet loaded when
-	 * MethodType gets loaded and this will cause the VM not to start.  See the code
-	 * in Class#getPrimitiveClass() for the issue. */
-	/*[ENDIF]*/
-	static final Class<?>[] primitivesArray = new Class<?>[26];
-	static {
-		primitivesArray['B' - 'A'] = byte.class;
-		primitivesArray['C' - 'A'] = char.class;
-		primitivesArray['D' - 'A'] = double.class;
-		primitivesArray['F' - 'A'] = float.class;
-		primitivesArray['I' - 'A'] = int.class;
-		primitivesArray['J' - 'A'] = long.class;
-		primitivesArray['S' - 'A'] = short.class;
-		primitivesArray['Z' - 'A'] = boolean.class;
-	}
-	
+
 	private static Map<MethodType, WeakReference<MethodType>> internTable = Collections.synchronizedMap(new WeakHashMap<MethodType, WeakReference<MethodType>>());
 	// lock object to control insertion of new MTs
 	static final class InternTableAddLock { }
@@ -135,7 +113,6 @@ public final class MethodType implements Serializable
 	private static final ObjectStreamField[] serialPersistentFields = new ObjectStreamField[0];
 	private static final long serialVersionUID = 292L;
 	private DeserializedFieldsHolder deserializedFields;
-
 
 	/**
 	 * Holder for the deserialized arguments[] and returnType values
@@ -227,7 +204,6 @@ public final class MethodType implements Serializable
 		}
 		return ints;
 	}
-
 
 	/**
 	 * Convenience method to create a new MethodType with only the parameter at 
@@ -410,32 +386,6 @@ public final class MethodType implements Serializable
 		
 		return methodType(returnType, types);
 	}
-	
-	/**
-	 * This helper calls MethodType.fromMethodDescriptorString(...) or 
-	 * MethodType.fromMethodDescriptorStringAppendArg(...) but throws 
-	 * NoClassDefFoundError instead of TypeNotPresentException during 
-	 * the VM resolve stage.
-	 *
-	 * @param methodDescriptor - the method descriptor string
-	 * @param loader - the ClassLoader to be used
-	 * @param appendArgumentType - an extra argument type
-	 *
-	 * @return a MethodType object representing the method descriptor string
-	 *
-	 * @throws IllegalArgumentException - if the string is not well-formed
-	 * @throws NoClassDefFoundError - if a named type cannot be found
-	 */
-	static final MethodType vmResolveFromMethodDescriptorString(String methodDescriptor, ClassLoader loader, Class<?> appendArgumentType) throws Throwable {
-		try {
-			if (null == appendArgumentType) {
-				return MethodType.fromMethodDescriptorString(methodDescriptor, loader);
-			}
-			return MethodType.fromMethodDescriptorStringAppendArg(methodDescriptor, loader, appendArgumentType);
-		} catch (TypeNotPresentException e) {
-			throw throwNoClassDefFoundError(e);
-		}
-	}
 
 	/**
 	 * Helper method to throw NoClassDefFoundError if the cause of TypeNotPresentException 
@@ -456,72 +406,6 @@ public final class MethodType implements Serializable
 			throw noClassDefFoundError;
 		}
 		throw e;
-	}
-	
-	/*
-	 * Convert the string from bytecode format to the format needed for ClassLoader#loadClass().
-	 * Change all '/' to '.'.
-	 * Remove the 'L' & ';' from objects, unless they are array classes.
-	 */
-	private static final Class<?> nonPrimitiveClassFromString(String name, ClassLoader classLoader) {
-		try {
-			name = name.replace('/', '.');
-			if (name.indexOf('L') == 0) {
-				// Remove the 'L' and ';'
-				name = name.substring(1, name.length() - 1);
-			}
-			return Class.forName(name, false, classLoader);
-		} catch(ClassNotFoundException e){
-			throw new TypeNotPresentException(name, e);
-		}
-	}
-
-	/*
-	 * Parses the first parameter in a descriptor string into a Class object.
-	 * The Class object is append to given ArrayList and the current string index is returned.
-	 */
-	static final int parseIntoClass(char[] signature, int index, ArrayList<Class<?>> args, ClassLoader classLoader, String descriptor) {
-		char current = signature[index];
-		Class<?> c;			
-		
-		if ((current == 'L') || (current == '[')) {
-			int start = index;
-			while(signature[index] == '[') {
-				index++;
-			}
-			String name;
-			if (signature[index] != 'L') {
-				name = descriptor.substring(start, index + 1);
-			} else { 
-				int end = descriptor.indexOf(';', index);
-				if (end == -1) {
-					/*[MSG "K05d6", "malformed method descriptor: {0}"]*/
-					throw new IllegalArgumentException(Msg.getString("K05d6", descriptor)); //$NON-NLS-1$
-				}
-				name = descriptor.substring(start, end + 1);
-				index = end;
-			}
-			c = nonPrimitiveClassFromString(name, classLoader);
-		} else {
-			try {
-				c = primitivesArray[current - 'A'];
-			} catch(ArrayIndexOutOfBoundsException e) {
-				c = null;
-			}
-			if (c == null) {
-				if (current == 'V') {
-					// lazily add 'V' to work around load ordering issues
-					primitivesArray['V' - 'A'] = void.class;
-					c = void.class;
-				} else {
-					/*[MSG "K05d7", "not a primitive: {0}"]*/
-					throw new IllegalArgumentException(Msg.getString("K05d7", current)); //$NON-NLS-1$
-				}
-			}
-		}
-		args.add(c);
-
-		return index;
 	}
 
 	/*
@@ -560,7 +444,7 @@ public final class MethodType implements Serializable
 				continue;
 			}
 
-			index = parseIntoClass(signature, index, args, classLoader, methodDescriptor);
+			index = MethodTypeHelper.parseIntoClass(signature, index, args, classLoader, methodDescriptor);
 			index++;
 		}
 		return args;
@@ -624,11 +508,11 @@ public final class MethodType implements Serializable
 	 * @return whether the MethodType contains any wrapper types
 	 */
 	public boolean hasWrappers() {
-		if (WRAPPER_SET.contains(returnType) || (returnType == Void.class)){
+		if (MethodTypeHelper.WRAPPER_SET.contains(returnType) || (returnType == Void.class)){
 			return true;
 		}
 		for(Class<?> c : arguments){
-			if (WRAPPER_SET.contains(c)){
+			if (MethodTypeHelper.WRAPPER_SET.contains(c)){
 				return true;
 			}
 		}
@@ -975,10 +859,10 @@ public final class MethodType implements Serializable
 	private String createMethodDescriptorString(){
 		StringBuilder sb= new StringBuilder("("); //$NON-NLS-1$
 		for (Class<?> c : arguments){
-			sb.append(getBytecodeStringName(c));
+			sb.append(MethodTypeHelper.getBytecodeStringName(c));
 		}
 		sb.append(")"); //$NON-NLS-1$
-		sb.append(getBytecodeStringName(returnType));
+		sb.append(MethodTypeHelper.getBytecodeStringName(returnType));
 		return sb.toString();
 	}
 	
@@ -1014,10 +898,10 @@ public final class MethodType implements Serializable
 	public MethodType unwrap(){
 		Class<?>[] args = arguments.clone();
 		for(int i = 0; i < args.length; i++){
-			args[i] = unwrapPrimitive(args[i]);
+			args[i] = MethodTypeHelper.unwrapPrimitive(args[i]);
 		}
-		Class<?> unwrappedReturnType = unwrapPrimitive(returnType);
-		if (unwrapPrimitive(returnType) == Void.class) {
+		Class<?> unwrappedReturnType = MethodTypeHelper.unwrapPrimitive(returnType);
+		if (MethodTypeHelper.unwrapPrimitive(returnType) == Void.class) {
 			unwrappedReturnType = void.class;
 		}
 		return methodType(unwrappedReturnType, args, false);
@@ -1034,9 +918,9 @@ public final class MethodType implements Serializable
 	public MethodType wrap(){
 		Class<?>[] args = arguments.clone();
 		for(int i = 0; i < args.length; i++){
-			args[i] = wrapPrimitive(args[i]);
+			args[i] = MethodTypeHelper.wrapPrimitive(args[i]);
 		}
-		return methodType(wrapPrimitive(returnType), args, false);
+		return methodType(MethodTypeHelper.wrapPrimitive(returnType), args, false);
 	}
 	
 	/**
@@ -1070,129 +954,7 @@ public final class MethodType implements Serializable
 		combinedParameters.addAll(classes);
 		return methodType(returnType, combinedParameters);
 	}
-	
-	/**
-	 * Returns the appropriate wrapper class or the original class
-	 * @param primitiveClass The class to convert to a wrapper class
-	 * @return The wrapper class or the original class if no wrapper is available
-	 */
-	static Class<?> wrapPrimitive(Class<?> primitveClass) {
-		if (primitveClass.isPrimitive()) {
-			if (int.class == primitveClass){ //I
-				return Integer.class;
-			}
-			if (long.class == primitveClass){ //J
-				return Long.class;
-			}
-			if (byte.class == primitveClass){	//B
-				return Byte.class;
-			}
-			if (char.class == primitveClass){ //C
-				return Character.class;
-			}
-			if (double.class == primitveClass){ //D
-				return Double.class;
-			}
-			if (float.class == primitveClass){ //F
-				return Float.class;
-			}
-			if (boolean.class == primitveClass){ //Z
-				return Boolean.class;
-			}
-			if (void.class == primitveClass) {
-				return Void.class;
-			}
-			if (short.class == primitveClass){ //S
-				return Short.class;
-			}
-		}
-		return primitveClass;
-	}
-	
-	
-	/**
-	 * Takes a class and returns the corresponding primitive if one exists.
-	 * @param wrapperClass The class to check for primitive classes.
-	 * @return The corresponding primitive class or the input class.
-	 */
-	/*[IF ]*/
-	/* Note that Void.class is not handled by this method as it is only viewed as a
-	 * wrapper when it is the return class.
-	 */
-	/*[ENDIF]*/
-	static Class<?> unwrapPrimitive(Class<?> wrapperClass){
-		if (Integer.class == wrapperClass) { //I
-			return int.class;
-		}
-		if (Long.class == wrapperClass) { //J
-			return long.class;
-		}
-		if (Byte.class == wrapperClass) {	//B
-			return byte.class;
-		}
-		if (Character.class == wrapperClass) { //C
-			return char.class;
-		}
-		if (Double.class == wrapperClass) { //D
-			return double.class;
-		}
-		if (Float.class == wrapperClass) { //F
-			return float.class;
-		}
-		if (Short.class == wrapperClass) { //S
-			return short.class;
-		}
-		if (Boolean.class == wrapperClass) { //Z
-			return boolean.class;
-		}
-		return wrapperClass;
-	}
-	
-	/* The string returned by this method should be in sync with Class.descriptorString() */
-	static String getBytecodeStringName(Class<?> c){
-		if (c.isPrimitive()) {
-			if (c == int.class) {
-				return "I"; //$NON-NLS-1$
-			} else if (c == long.class) {
-				return "J"; //$NON-NLS-1$
-			} else if (c == byte.class) {
-				return "B"; //$NON-NLS-1$
-			} else if (c == boolean.class) {
-				return "Z"; //$NON-NLS-1$
-			} else if (c == void.class) {
-				return "V"; //$NON-NLS-1$
-			} else if (c == char.class) {
-				return "C"; //$NON-NLS-1$
-			} else if (c == double.class) {
-				return "D"; //$NON-NLS-1$
-			} else if (c == float.class) {
-				return "F"; //$NON-NLS-1$
-			} else if (c == short.class) {
-				return "S"; //$NON-NLS-1$
-			}
-		}
-		Class<?> clazz = c;
-		if (c.isArray()) {
-			clazz = c.getComponentType();
-			while (clazz.isArray()) {
-				clazz = clazz.getComponentType();
-			}
-		}
-		String name = c.getName().replace('.', '/');
-		/*[IF Java15]*/
-		if (clazz.isHidden()) {
-			/* keep the last "." before romaddress for hidden classes */
-			int index = name.lastIndexOf('/');
-			name = name.substring(0, index) + '.' + name.substring(index + 1,name.length());
-		}
-		/*[ENDIF] Java15 */
-		
-		if (c.isArray()) {
-			return name;
-		}
-		return "L"+ name + ";"; //$NON-NLS-1$ //$NON-NLS-2$
-	}
-	
+
 	/*
 	 * Implements JSR 292 serialization spec.
 	 * Write the MethodType without writing the field names.
